ENH:  Move history to python module
