+++ OsteotomyPlanner/OsteotomyPlanner.py
@@ -4,6 +4,7 @@ import logging
 import vtk, qt, ctk, slicer
 from slicer.ScriptedLoadableModule import *
 from slicer.util import VTKObservationMixin
+import ModelHistory.ModelHistory as ModelHistory
 
 #
 # OsteotomyPlanner
@@ -139,6 +140,8 @@ class OsteotomyPlannerWidget(ScriptedLoadableModuleWidget, VTKObservationMixin):
     self.ui.SubjectHierarchyTreeView.hideColumn(4)
     self.ui.SubjectHierarchyTreeView.hideColumn(5)
 
+
+
     #Selection 
     self.ui.SubjectHierarchyTreeView.setSelectionMode(qt.QAbstractItemView().SingleSelection)
     self.ui.ActionsWidget.setCurrentWidget(self.ui.MenuWidget)
@@ -164,169 +167,40 @@ class OsteotomyPlannerWidget(ScriptedLoadableModuleWidget, VTKObservationMixin):
     self.activeNode = None
     self.splitPlanes = []
     self.splitModels = []
-    self.history = [] #List of states, each state is a list of nodes
-    self.maximumSavedStates = 10
-    self.lastRestoredState = 0 
-    self.cachedState = None
+    # self.history = [] #List of states, each state is a list of nodes
+    # self.maximumSavedStates = 10
+    # self.lastRestoredState = 0 
+    # self.cachedState = None
     self.actionInProgress = False
-    
+    self.modelHistory = ModelHistory()    
 
     # Make sure parameter node is initialized (needed for module reload)
     self.initializeParameterNode()
     self.resolveStateButtons()
 
+  def restoreNextState(self):
+    self.modelHistory.restoreNextState()
+    self.resolveStateButtons()
+
+  
+  def restorePreviousState(self):
+    self.modelHistory.restorePreviousState()
+    self.resolveStateButtons()
+
+  
   def finishPlan(self):
-    self.clearHistory()
+    self.modelHistory.clearHistory()
     self.resolveStateButtons()
     self.ui.SubjectHierarchyComboBox.setCurrentItem(-1)
-    self.ui.SubjectHierarchyTreeView.enabled = False
-  
-  def archiveNode(self, node):
-    node.HideFromEditorsOn()
-    node.SetDisplayVisibility(False)
-    node.SetAttribute("History", "yes")
-
-  def restoreNode(self, node):
-    node.HideFromEditorsOff()
-    node.SetDisplayVisibility(True)
-    node.RemoveAttribute("History")
-
-  def removeNode(self,node):
-    slicer.mrmlScene.RemoveNode(node)
-
-  def isNodeCurrent(self, node):
-    return not node.GetAttribute("History") == "yes"
-  
-  def saveState(self):
-    if self.maximumSavedStates < 1:
-      return
-    # self.removeAllNextStates()
-    if self.cachedState is None:
-      self.cacheState()
-    self.history.append(self.cachedState)
-    self.cachedState = None
-    self.lastRestoredState = len(self.history)
-    self.removeAllObsoleteStates()
-
-  
-  def cacheState(self):
-    self.clearCachedState()
-    self.cachedState = []
-    #clone nodes
-    children = vtk.vtkIdList()
-    shNode = slicer.vtkMRMLSubjectHierarchyNode.GetSubjectHierarchyNode(slicer.mrmlScene)
-
-    shNode.GetItemChildren(self.ui.SubjectHierarchyComboBox.currentItem(), children) # Add a third argument with value True for recursive query
-    state = []
-    for i in range(children.GetNumberOfIds()):
-      child = children.GetId(i)
-      childNode = shNode.GetItemDataNode(child)
-      if not self.isNodeCurrent(childNode):
-        continue
-      clonedChild = self.cloneNode(childNode)
-      self.archiveNode(clonedChild)
-      self.cachedState.append(clonedChild)
-  
-  def clearCachedState(self):
-    if self.cachedState is not None:
-      self.deleteState(self.cachedState)
-    self.cachedState = None
-  
-  def removeAllNextStates(self):
-    while (len(self.history) > (self.lastRestoredState + 1)) and self.history:
-      state = self.history.pop()
-      self.deleteState(state)
-    
-
-  def removeAllObsoleteStates(self):
-    while (len(self.history) > self.maximumSavedStates) and self.history:
-      state = self.history.pop(0)
-      self.lastRestoredState -= 1
-      self.deleteState(state)
-  
-  def deleteState(self, state):
-    for node in state:
-      self.removeNode(node)
-
-  def restoreState(self, stateIndex):
-    #restore state, then update index
-    restoredState = self.history.pop()
-    
-    #drop all of the current nodes
-    children = vtk.vtkIdList()
-    shNode = slicer.vtkMRMLSubjectHierarchyNode.GetSubjectHierarchyNode(slicer.mrmlScene)
-
-    shNode.GetItemChildren(self.ui.SubjectHierarchyComboBox.currentItem(), children) # Add a third argument with value True for recursive query
-    for i in range(children.GetNumberOfIds()):
-      child = children.GetId(i)
-      childNode = shNode.GetItemDataNode(child)
-      if self.isNodeCurrent(childNode):
-        self.removeNode(childNode)
-
-    for model in restoredState:
-      self.restoreNode(model)
-
-    self.lastRestoredState = stateIndex
-
-  def restorePreviousState(self):
-    if self.lastRestoredState < 1:
-      print("There are no previous state available for restore")   
-      return
-
-    if len(self.history) < self.lastRestoredState:
-      print("There are no previous state available for restore")   
-      return
-
-    stateToRestore = self.lastRestoredState -1
-    self.restoreState(stateToRestore)
-    self.resolveStateButtons()
-
-  def restoreNextState(self):
-
-    if self.lastRestoredState + 1 >= len(self.history):
-      print("No next state available to restore")
-      return
-
-    self.restoreState(self.lastRestoredState + 1)
-    self.resolveStateButtons()
+    self.ui.SubjectHierarchyTreeView.enabled = False  
 
   def resolveStateButtons(self):
-
-    self.ui.UndoButton.enabled = self.isRestorePreviousStateAvailable() and not self.actionInProgress
-    self.ui.RedoButton.enabled = self.isRestoreNextStateAvailable() and not self.actionInProgress
-    self.ui.SubjectHierarchyComboBox.enabled = (len(self.history) == 0)
+    self.ui.UndoButton.enabled = self.modelHistory.isRestorePreviousStateAvailable() and not self.actionInProgress
+    self.ui.RedoButton.enabled = self.modelHistory.isRestoreNextStateAvailable() and not self.actionInProgress
+    self.ui.SubjectHierarchyComboBox.enabled = not self.modelHistory.hasHistory() and not self.actionInProgress
     self.ui.FinishButton.enabled = not self.actionInProgress
     # print(self.history)
-    # print("LastRestored: " + str(self.lastRestoredState))
-
-  
-  def isRestorePreviousStateAvailable(self):
-    return not (self.lastRestoredState < 1)
-
-  def isRestoreNextStateAvailable(self):
-    # return not (self.lastRestoredState + 1 >= len(self.history))
-    return False
-  
-  
-  def clearHistory(self):
-    for state in self.history:
-      self.deleteState(state)
-
-    self.history = []
-    self.lastRestoredState = 0
-
-  def getNumberOfStates(self):
-    return len(self.history)
-  
-  def cloneNode(self, nodeToClone):
-    # Clone the node
-    shNode = slicer.vtkMRMLSubjectHierarchyNode.GetSubjectHierarchyNode(slicer.mrmlScene)
-    itemIDToClone = shNode.GetItemByDataNode(nodeToClone)
-    clonedItemID = slicer.modules.subjecthierarchy.logic().CloneSubjectHierarchyItem(shNode, itemIDToClone)
-    clonedNode = shNode.GetItemDataNode(clonedItemID)
-    clonedNode.SetName(nodeToClone.GetName())
-    return clonedNode
-  
+    # print("LastRestored: " + str(self.lastRestoredState))  
   
   def placeManualPlane(self):
     interactionNode = slicer.app.applicationLogic().GetInteractionNode()
@@ -393,7 +267,7 @@ class OsteotomyPlannerWidget(ScriptedLoadableModuleWidget, VTKObservationMixin):
     self.removeNode(self.activeNode)
     #need to force selection of something
     self.ui.SubjectHierarchyTreeView.setCurrentItem(selectItem)
-    self.saveState()
+    self.modelHistory.saveState()
     self.endSplit()
 
   def clearSplitModels(self):
@@ -409,7 +283,7 @@ class OsteotomyPlannerWidget(ScriptedLoadableModuleWidget, VTKObservationMixin):
   def beginAction(self):
     self.ui.SubjectHierarchyTreeView.setSelectionMode(qt.QAbstractItemView().NoSelection)
     self.ui.MenuWidget.enabled = False
-    self.cacheState()
+    self.modelHistory.cacheState()
     self.actionInProgress = True
     self.resolveStateButtons()
 
@@ -440,7 +314,7 @@ class OsteotomyPlannerWidget(ScriptedLoadableModuleWidget, VTKObservationMixin):
   def confirmMove(self):
     logic = slicer.vtkSlicerTransformLogic()
     logic.hardenTransform(self.activeNode)
-    self.saveState()
+    self.modelHistory.saveState()
     self.endMove()
 
   def endSplit(self):
@@ -470,7 +344,7 @@ class OsteotomyPlannerWidget(ScriptedLoadableModuleWidget, VTKObservationMixin):
     self.ui.SubjectHierarchyTreeView.setRootItem(item)
     self.ui.SubjectHierarchyTreeView.expandItem(item)
     self.ui.SubjectHierarchyTreeView.enabled = True
-    self.clearHistory()
+    self.modelHistory = ModelHistory(item)
 
   
   def cleanup(self):
